# backend/src/main/java/com/bankapp/auth/api/LoginRequest.java
package com.bankapp.auth.api;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

public record LoginRequest(
        @NotBlank(message = "Email is required.")
        @Email(message = "Email format is invalid.")
        String email,
        @NotBlank(message = "Password is required.")
        String password
) {
}