# specs/001-bank-auth-customer/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Bank App Identity and Customer Management API
  version: 0.1.0
  description: OpenAPI-first skeleton for authentication and customer management.
servers:
  - url: http://localhost:8080
tags:
  - name: Auth
  - name: Customers
paths:
  /api/auth/register:
    post:
      tags: [Auth]
      operationId: registerUser
      summary: Register a new user account.
      description: Create a new active user with a unique email address and a BCrypt-hashed password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              valid:
                value:
                  email: customer@example.com
                  password: SecurePass1!
      responses:
        '201':
          description: User registered successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              examples:
                created:
                  value:
                    userId: 86cbf1e1-f07a-4e46-b666-1ec8f6f47128
                    email: customer@example.com
                    roles: [CUSTOMER]
                    isActive: true
                    createdAt: 2026-04-08T10:15:30Z
                    updatedAt: 2026-04-08T10:15:30Z
        '409':
          description: Duplicate user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                duplicate:
                  value:
                    code: USER_ALREADY_EXISTS
                    message: A user with this email already exists.
                    field: email
        '422':
          description: Validation failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidPassword:
                  value:
                    code: INVALID_PASSWORD_FORMAT
                    message: Password must be at least 8 characters and include an uppercase letter, digit, and special character.
                    field: password
  /api/auth/login:
    post:
      tags: [Auth]
      operationId: loginUser
      summary: Authenticate an existing user.
      description: Log in an active user and return signed access and refresh tokens.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              valid:
                value:
                  email: customer@example.com
                  password: SecurePass1!
      responses:
        '200':
          description: Authenticated session response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
              examples:
                ok:
                  value:
                    accessToken: access.jwt.token
                    refreshToken: refresh.jwt.token
                    tokenType: Bearer
                    expiresIn: 3600
        '401':
          description: Invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidCredentials:
                  value:
                    code: INVALID_CREDENTIALS
                    message: The supplied credentials are invalid.
        '403':
          description: Inactive account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                inactive:
                  value:
                    code: ACCOUNT_INACTIVE
                    message: The account is inactive.
  /api/customers:
    post:
      tags: [Customers]
      operationId: createCustomer
      security:
        - bearerAuth: []
      summary: Create a customer profile.
      description: Create a PERSON or COMPANY customer linked to the authenticated user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
            examples:
              person:
                value:
                  name: Jamie Customer
                  address: 10 Main Street
                  type: PERSON
      responses:
        '201':
          description: Customer created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
              examples:
                created:
                  value:
                    customerId: 1001
                    name: Jamie Customer
                    address: 10 Main Street
                    type: PERSON
                    createdAt: 2026-04-08T10:15:30Z
                    updatedAt: 2026-04-08T10:15:30Z
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                unauthorised:
                  value:
                    code: UNAUTHORISED
                    message: Access is not authorised for this request.
        '422':
          description: Validation failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidType:
                  value:
                    code: INVALID_CUSTOMER_TYPE
                    message: Customer type is invalid.
                    field: type
  /api/customers/{customerId}:
    parameters:
      - $ref: '#/components/parameters/CustomerId'
    patch:
      tags: [Customers]
      operationId: updateCustomer
      security:
        - bearerAuth: []
      summary: Update mutable customer fields.
      description: Update mutable customer fields while rejecting immutable or stale requests.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerRequest'
            examples:
              valid:
                value:
                  name: Updated Customer
                  address: 20 Main Street
                  type: COMPANY
                  updatedAt: 2026-04-08T10:15:30Z
      responses:
        '200':
          description: Customer updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          description: Immutable or malformed field supplied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                immutable:
                  value:
                    code: FIELD_NOT_UPDATABLE
                    message: Immutable customer fields cannot be updated.
                    field: email
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Customer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Stale update conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                stale:
                  value:
                    code: CUSTOMER_CONFLICT
                    message: Customer state is stale. Refresh and try again.
    get:
      tags: [Customers]
      operationId: getCustomer
      security:
        - bearerAuth: []
      summary: Retrieve customer details.
      description: Retrieve a customer owned by the authenticated user.
      responses:
        '200':
          description: Customer detail response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          description: Invalid identifier.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Customer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CustomerId:
      name: customerId
      in: path
      required: true
      schema:
        type: integer
        format: int64
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        field:
          type: string
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          pattern: ^(?=.*[A-Z])(?=.*\d)(?=.*[^A-Za-z0-9]).{8,}$
    UserResponse:
      type: object
      required: [userId, email, roles, isActive, createdAt, updatedAt]
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        roles:
          type: array
          items:
            type: string
            enum: [CUSTOMER, ADMIN]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    SessionResponse:
      type: object
      required: [accessToken, refreshToken, tokenType, expiresIn]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          format: int64
    CreateCustomerRequest:
      type: object
      required: [name, address, type]
      properties:
        name:
          type: string
          minLength: 2
        address:
          type: string
        type:
          $ref: '#/components/schemas/CustomerType'
    UpdateCustomerRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        address:
          type: string
        type:
          $ref: '#/components/schemas/CustomerType'
        email:
          type: string
          format: email
        accountNumber:
          type: string
        updatedAt:
          type: string
          format: date-time
    CustomerResponse:
      type: object
      required: [customerId, name, address, type, createdAt, updatedAt]
      properties:
        customerId:
          type: integer
          format: int64
        name:
          type: string
        address:
          type: string
        type:
          $ref: '#/components/schemas/CustomerType'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CustomerType:
      type: string
      enum: [PERSON, COMPANY]